backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
langgraph==0.2.34
langchain-openai==0.2.8
langchain-core==0.3.75
pydantic==2.9.2
python-dotenv==1.0.0